backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
pgvector==0.3.2
numpy==1.26.4
opencv-python-headless==4.10.0.84
mediapipe==0.10.18
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.8.2
google-generativeai==0.7.2
httpx==0.27.0
